+++ core/src/main/scala/dogs/List.scala
@@ -451,7 +451,7 @@ object List extends ListInstances {
   }
 
 
-  implicit def toPartition[A](aList: List[A]): Partition[A] = Partition(aList)
+  implicit def toPartition[A](aList: List[A]): BiPartition[A] = BiPartition(aList)
 
   implicit def toSorted[A](a: List[A])(implicit order: Order[A]): Sorted[A] = Sorted(a)
